+++ src/aplicaciones/sistemasNumericos/helpers/sistemasNumericos/Octal.java
@@ -17,13 +17,13 @@ public class Octal extends SistemaNumerico{
 	}
 	
 	public String ADecimal() {
-		int resultado=0;
-		int num=Integer.parseInt(this.getValue());
+		long resultado=0;
+		long num=Long.parseLong(this.getValue());
 		for(int i=0;i<this.getValue().length();i++) {
-			resultado=(int) (num%10*Math.pow(this.getBase(), i)+resultado);
+			resultado=(long) (num%10*Math.pow(this.getBase(), i)+resultado);
 			num=num/10;
 		}
-		return Integer.toString(resultado);
+		return Long.toString(resultado);
 	}
 	
 	public String ABinario() {
